--- Training_node/first_project/app/controllers/albumes.js
@@ -1,13 +1,0 @@
-const errors = require('../errors'),
-  request = require('request'),
-  logger = require('../logger');
-
-exports.getAll = (req, res, next) => {
-  return request('https://jsonplaceholder.typicode.com/albums', function(error, response, body) {
-    if (!error && response.statusCode === 200) {
-      const info = JSON.parse(body);
-      // do more stuff
-      res.send({ albumes: info });
-    }
-  });
-};
