+++ Training_node/first_project/app/routes.js
@@ -1,5 +1,5 @@
 const users = require('./controllers/users'),
-  albumes = require('./controllers/albumes'),
+  albums = require('./controllers/albums'),
   auth = require('./middlewares/auth');
 
 exports.init = app => {
@@ -8,5 +8,5 @@ exports.init = app => {
   app.post('/users/sessions', [], users.login);
   app.get('/users', [auth.secureRegular], users.getAll);
   app.post('/admin/users', [auth.secureAdmin], users.createAdmin);
-  app.get('/albumes', [auth.secureRegular], albumes.getAll);
+  app.get('/albums', [auth.secureRegular], albums.getAll);
 };
