+++ Training_node/first_project/config/index.js
@@ -48,6 +48,10 @@ const config = {
     rollbar: {
       accessToken: process.env.ROLLBAR_ACCESS_TOKEN,
       environment: process.env.ROLLBAR_ENV
+    },
+    urlRequests: {
+      base: 'https://jsonplaceholder.typicode.com',
+      albumList: '/albums'
     }
   }
 };
