+++ Training_node/first_project/docs/users.html
@@ -12,10 +12,10 @@
   "firstName": "julian",
   "lastName": "molina",
   "email": "julian.molina@wolox.com.ar",
-  "password": "$2a$10$r57x0Fxg3dm9UDrV7P0KXeW62nebLorNzokiPvCoBBJTeHuT2jdDy",
+  "password": "$2a$10$V1gHyLAxgFUrfjVOSzL57.bnQcTNaJIssMzXKJivGNYESXMUtxOFq",
   "admin": true,
-  "created_at": "2018-05-21T11:27:05.033Z",
-  "updated_at": "2018-05-21T11:27:05.033Z",
+  "created_at": "2018-05-21T14:35:26.524Z",
+  "updated_at": "2018-05-21T14:35:26.524Z",
   "deleted_at": null
 }</pre></div></div></div></div><div class="panel panel-default"><div class="panel-heading" role="tab" id="heading-users-POST"><h4 class="panel-title"><a role="button" data-toggle="collapse" data-parent="#accordion" aria-expanded="true" aria-controls="collapse-users-POST" href="#collapse-users-POST">POST /users</a></h4></div><div id="collapse-users-POST" class="panel-collapse collapse" role="tabpanel" aria-labelledby="heading-users-POST"><div class="panel-body"><h4>Request Data</h4><div class="container"><h5>Headers</h5><pre>{
   "content-type": "application/json",
@@ -39,10 +39,10 @@
       "firstName": "julian",
       "lastName": "sevilla",
       "email": "julian.sevilla@wolox.com.ar",
-      "password": "$2a$10$Cdag9rCT7uKpgPiSEBZexeM6wR5ZtA/xHRdjdv7827LUSjdLM0uDu",
+      "password": "$2a$10$VYzbxsilSpnMLevpzFlupOng/IV0k2LW1RKin/XqSylDJjrkeGHtO",
       "admin": null,
-      "created_at": "2018-05-21T11:27:05.699Z",
-      "updated_at": "2018-05-21T11:27:05.699Z",
+      "created_at": "2018-05-21T14:35:27.149Z",
+      "updated_at": "2018-05-21T14:35:27.149Z",
       "deleted_at": null
     },
     {
@@ -50,10 +50,10 @@
       "firstName": "pablo",
       "lastName": "cid",
       "email": "pablo.cid@wolox.com.ar",
-      "password": "$2a$10$Cdag9rCT7uKpgPiSEBZexeM6wR5ZtA/xHRdjdv7827LUSjdLM0uDu",
+      "password": "$2a$10$VYzbxsilSpnMLevpzFlupOng/IV0k2LW1RKin/XqSylDJjrkeGHtO",
       "admin": null,
-      "created_at": "2018-05-21T11:27:05.699Z",
-      "updated_at": "2018-05-21T11:27:05.699Z",
+      "created_at": "2018-05-21T14:35:27.149Z",
+      "updated_at": "2018-05-21T14:35:27.149Z",
       "deleted_at": null
     }
   ]
@@ -68,10 +68,10 @@
       "firstName": "pablo",
       "lastName": "cid",
       "email": "pablo.cid@wolox.com.ar",
-      "password": "$2a$10$6oGvfnXWFlt7yrsOOCrVLODOijUeboHXYz3kIVUDrtP38tODHVN4m",
+      "password": "$2a$10$Od..kJFqKlLmSTYfFe0WW.LeqwCK7fjmReQCAY6gvDjlSqHCKZSVe",
       "admin": null,
-      "created_at": "2018-05-21T11:27:05.857Z",
-      "updated_at": "2018-05-21T11:27:05.857Z",
+      "created_at": "2018-05-21T14:35:27.343Z",
+      "updated_at": "2018-05-21T14:35:27.343Z",
       "deleted_at": null
     }
   ]
