+++ Training_node/first_project/package.json
@@ -44,6 +44,7 @@
     "express": "^4.16.2",
     "jwt-simple": "^0.5.1",
     "morgan": "^1.9.0",
+    "nock": "^9.2.6",
     "pg": "^7.4.1",
     "pg-native": "^2.2.0",
     "request": "^2.86.0",
